refactor: изменен путь к исходникам.
