+++ src/rest/connector/ConnectorBase.ts
